# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Prélèvements d'eau API
  version: 1.0.0
  description: API HTTP de Partageons l'eau pour les déclarations, référentiels, séries temporelles et imports.
servers:
  - url: http://localhost:5000/api
    description: Local
...
paths:
...
  /auth/request:
    post:
      tags: [Auth]
      operationId: requestAuth
      summary: Demander un lien de connexion
      description: Envoie un jeton de connexion à l'adresse email fournie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: {type: string, format: email}
      responses:
        '204': {description: Demande acceptée}
        default: {$ref: '#/components/responses/Error'}
...
  /auth/verify:
    post:
      tags: [Auth]
      operationId: verifyAuthToken
      summary: Vérifier un jeton de connexion
      description: Crée une session à partir d'un jeton d'authentification valide.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
      responses:
        '200':
          description: Session créée
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResponse'}
        default: {$ref: '#/components/responses/Error'}
...
  /auth/password:
    post:
      tags: [Auth]
      operationId: authenticateWithPassword
      summary: Se connecter par email et mot de passe
      description: Cette route existe uniquement lorsque la méthode `password` est active.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: {type: string, format: email}
                password: {type: string, minLength: 1, maxLength: 128, writeOnly: true}
      responses:
        '200':
          description: Session créée
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResponse'}
        '401': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        default: {$ref: '#/components/responses/Error'}
  /auth/password/activate:
    post:
      tags: [Auth]
      operationId: activatePassword
      summary: Choisir le mot de passe initial
      description: Consomme un lien d’activation à usage unique valable 72 heures.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, password]
              properties:
                token: {type: string, writeOnly: true}
                password: {type: string, minLength: 15, maxLength: 128, writeOnly: true}
      responses:
        '200':
          description: Mot de passe activé et session créée
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResponse'}
        '429': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        default: {$ref: '#/components/responses/Error'}
  /auth/password/change:
    post:
      tags: [Auth]
      operationId: changePassword
      summary: Modifier son mot de passe
      description: Invalide les autres sessions et retourne une nouvelle session.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currentPassword, newPassword]
              properties:
                currentPassword: {type: string, minLength: 1, maxLength: 128, writeOnly: true}
                newPassword: {type: string, minLength: 15, maxLength: 128, writeOnly: true}
      responses:
        '200':
          description: Mot de passe modifié et session remplacée
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthSessionResponse'}
        default: {$ref: '#/components/responses/Error'}
...
  /users/me/email-aliases:
    get:
      tags: [Compte]
      operationId: listMyEmailAliases
      summary: Lister mes autres adresses de connexion validées
      description: Retourne uniquement les adresses déjà validées et utilisables pour se connecter.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Adresses validées
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [emailAliases]
                properties:
                  emailAliases:
                    type: array
                    items: {$ref: '#/components/schemas/UserEmailAlias'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Compte]
      operationId: requestEmailAliasAddition
      summary: Demander l’ajout d’une autre adresse de connexion
      description: L’adresse devient utilisable uniquement après validation du lien reçu.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EmailVerificationRequest'}
      responses:
        '202':
          description: Demande enregistrée; le statut indique si l’email a pu être envoyé
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserEmailVerification'}
        '409': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
    Error:
      description: Erreur
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Uuid:
      type: string
      format: uuid
    Error:
      type: object
      required: [message]
      properties:
        message: {type: string}
        code: {type: string}
        details: {}
...
    UserEmailAlias:
      type: object
      additionalProperties: false
      required: [id, email, createdAt, updatedAt]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        email: {type: string, format: email}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    UserEmailVerification:
      type: object
      additionalProperties: false
      required: [id, purpose, email, status, createdAt, sentAt, expiresAt, verifiedAt, cancelledAt, nextResendAt, canResend]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        purpose: {type: string, enum: [PRIMARY_CHANGE, ALIAS_ADD]}
        email: {type: string, format: email}
        status:
          type: string
          enum: [PENDING, SEND_FAILED, EXPIRED, VERIFIED, CANCELLED, SUPERSEDED, CONFLICT]
        createdAt: {type: string, format: date-time}
        sentAt: {type: [string, 'null'], format: date-time}
        expiresAt: {type: string, format: date-time}
        verifiedAt: {type: [string, 'null'], format: date-time}
        cancelledAt: {type: [string, 'null'], format: date-time}
        nextResendAt: {type: [string, 'null'], format: date-time}
        canResend: {type: boolean}
    EmailVerificationRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email: {type: string, format: email, maxLength: 320}
...
    AuthSessionResponse:
      type: object
      additionalProperties: false
      required: [success, token, expiresAt]
      properties:
        success: {type: boolean, const: true}
        token: {type: string, writeOnly: true}
        expiresAt: {type: string, format: date-time}
...